删除数据前增加判断，增加title,name的数据验证

+++ application/api/controller/Ad.php
@@ -179,12 +179,14 @@ class Ad extends Base
     public function delete($id)
     {
         $ad = AdModel::get($id);
-       
         if (!$ad) {
             ajax_return(ResultCode::E_DATA_NOT_FOUND, '广告不存在!');
         }
 
-        $ad->delete();
+        $res = $ad->delete();
+        if (!$res) {
+            return ajax_return(ResultCode::E_DB_ERROR, '操作失败!');
+        }
 
         $AdServingModel = new AdServingModel();
         $AdServingModel->where('ad_id', $id)->delete();
